Move the preprocessor functions from app.py to utils.py
