+++ flask_app/app.py
@@ -1,12 +1,7 @@
 from flask import Flask, redirect, url_for, render_template, request
 import pickle
-import nltk
-from nltk.corpus import stopwords 
-from nltk.corpus import wordnet
-from nltk.stem import WordNetLemmatizer
-import string
 
-from utils import sentence_preprocessor
+from utils import sentence_preprocessor, get_topics
 
 app = Flask(__name__)
 
@@ -24,52 +19,23 @@ def try_out():
         text = request.form.get('text')
 
         ##### NER Model #####
-
         df_list, intermediate_df = sentence_preprocessor(text)
-
         with open('ner_model.pkl', 'rb') as ner_f:
-            pipeline = pickle.load(ner_f)
-        preds = pipeline.predict(intermediate_df)
-
+            ner_model = pickle.load(ner_f)
+        preds = ner_model.predict(intermediate_df)
         rtn_ners = []
         for pred, word in zip(preds, df_list):
             if pred == 1:
                 rtn_ners.append(word[0])
-        
             ## add spacy ner model here. 
 
         ##### LDA Model #####
-        dct = pickle.load(open('dct.pkl', 'rb')) 
+        with open('dct.pkl', 'rb') as dct_f:
+            dictionary = pickle.load(dct_f)
+        with open('lda.pkl', 'rb') as lda_f:
+            lda_model = pickle.load(lda_f)
         lda_model = pickle.load(open('lda.pkl', 'rb'))   
-
-        def _lemmatize_words(sentence):
-            wordnet_map = {'N':wordnet.NOUN, 'V':wordnet.VERB, 'J':wordnet.ADJ, 'R':wordnet.ADV}
-            pos_tagged_text = nltk.pos_tag(sentence.split())
-            return ' '.join([WordNetLemmatizer().lemmatize(word, wordnet_map.get(pos[0], wordnet.NOUN)) for word, pos in pos_tagged_text])
-    
-        def lda_sent_process(text):
-            text = text.lower()  
-            PUNCT_TO_REMOVE = string.punctuation
-            text = text.translate(str.maketrans('', '', PUNCT_TO_REMOVE))
-            STOPWORDS = set(stopwords.words('english'))
-            text = ' '.join([word for word in text.split() if word not in STOPWORDS])
-            text = _lemmatize_words(text)
-            return text.split()
-
-        def get_topics(new_text, lda_model, dct): 
-            '''
-            new_text: str
-            lda_model: load from lda.pkl
-            dct: load from dct.pkl
-            '''
-            rtn_list = []
-            new_text_doc = lda_sent_process(new_text)
-            topics = lda_model[dct.doc2bow(new_text_doc)]
-            for topic in topics: 
-                rtn_list.append(f'Topic {topic[0]} with probability {topic[1]}')
-            return rtn_list
-
-        lda_text = get_topics(text, lda_model, dct)
+        lda_text = get_topics(text, lda_model, dictionary)
     
         return render_template('try.html', requested=True, ner_text=rtn_ners, lda_text=lda_text)
     return render_template('try.html', requested=False)
